+++ src/types/standing.ts
@@ -9,4 +9,5 @@ export interface Standing {
   buchholz: number
   medianBuchholz: number
   sonnebornBerger: number
+  dropped: boolean
 }